Changed the aiming of the bullet and a few other things.
Added a very simple way of doing the animation

The animation is played at 7 fps because I thought it looked decent.

The projectile now shoots in the direction the mouse cursor is in.

This is a much better way than how we were doing things, but it is in the experiment branch so people can tell me if they like it. Let me know on discord.